drivers: acpi: Add MSI support to evged
The evged driver was only able to support legacy interrupts coming
from the IOAPIC. This patch allows this same driver to handle MSI
messages, for platforms looking at legacy free interrupts.

Note that the geddev structure is passed through the convenient void
pointer that was previously added to irq_alloc_info structure. This
allows the msi_write() callback to retrieve the information about the
device, which is used to call into the _MSI ACPI method.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>