irq: Create a new attribute to pass any kind of information
The structure irq_alloc_info is mostly attributes defined by specific
configurations. As we don't want to define yet another configuration,
defining a generic void pointer as part of this structure is very
convenient. Any user of this structure can now pass any type of data
to be retrieved later.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>