ACPI/IORT: Pass the pci bus handle to rc_dma_get_range
We currently pass the end point struct device handle to
rc_dma_get_range(). This handle then is passed as a context to
the iort_match_node_callback() callback when executing
iort_scan_node, looking for the RC node. iort_match_node_callback
expects a pci_bus device handle instead as to_pci_bus is called.
Then it may crash the kernel when doing the pci_domain_nr().

---

The problem was observed when using the IORT code on x86 for the
virtio-iommu integration

Signed-off-by: Eric Auger <eric.auger@redhat.com>