iommu/virtio: HACK: Use the fwnode allocated by the iort code
When integration the virtio-iommu on x86; I noticed the
dev->fwnode of the virtio-iommu-pci device already is populated
by the ACPI glue so the FW node attached to the iommu_device through
iommu_device_set_fwnode does not match the one allocated by the
IORT code. So the iommu setup fails.

Signed-off-by: Eric Auger <eric.auger@redhat.com>

---

This is a temporary hack.