iommu/virtio: Support topology description from virtio-mmio based virtio-iommu
Support topology description from virtio-iommu that's based on the
virtio-mmio transport.

We could actually call the topology parser directly from viommu_probe().
Indeed, on a platform without ACPI or DT a virtio-mmio device can only
be instantiated using the "device" module parameter. So the device is
created by the virtio-mmio module itself, and there is no need to use a
bus notifier. It would be more efficient but I'm still not sure it works
in all cases.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>